Capa de Acceso a datos Se modifico y avanazo Interfaces , Servicios , Accesos para Docentes y Sentencias para el INSERT de estos Se arreglo enum para ser mostrados en datagridviews Se añadio mas Scripts SQL para INSERTS

+++ UGELNorte.Resoluciones.Core/Models/DocenteModel.cs
@@ -7,7 +7,7 @@ namespace UGELNorte.Resoluciones.Core.Models
 {
     public class DocenteModel
     {
-        public int DNI { get; set; }
+        public Int64 DNI { get; set; }
 
         public string apellidoPaterno { get; set; }
